Add files via upload
This project consisting a flight booking landing page. The page  features a stylish and responsive design with a navigation bar containing sections like Apps, DAO, Technology, Token, and Contributors. Users are encouraged to take action by buying tokens, joining the DAO, and contributing to positive change. The hero section highlights the project's mission with a compelling message, accompanied by an option to request early access. The visual elements are carefully crafted, utilizing a dark theme for a modern and immersive user experience. The project's code includes HTML and CSS to structure and style the web pages effectively.